% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSparkRConnection.R
\name{createSparkRConnection}
\alias{createSparkRConnection}
\title{createSparkRConnection
Return an instance of the SparkRConnection from a Spark Context object.}
\usage{
createSparkRConnection(sc, ...)
}
\arguments{
\item{sc}{Spark Context object}

\item{...}{Extra parameters}
}
\description{
createSparkRConnection
Return an instance of the SparkRConnection from a Spark Context object.
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
dbGetQuery(db, "SELECT * FROM mtcars WHERE cyl == 4")
}
}
