% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbBind.R
\name{dbBind,SparkRResult-method}
\alias{dbBind,SparkRResult-method}
\title{dbBind DBI method
Parametrised query is not officialy supported in Spark environment,
it is here managed with the sqlInterpolate method
DBI documentation: https://dbi.r-dbi.org/reference/dbBind.html}
\usage{
\S4method{dbBind}{SparkRResult}(res, params, ...)

\S4method{dbBind}{SparkRResult}(res, params, ...)
}
\arguments{
\item{res}{SparkRResult object}

\item{params}{List of parameters to fill the parametrised query}

\item{...}{Extra parameters}
}
\description{
dbBind DBI method
Parametrised query is not officialy supported in Spark environment,
it is here managed with the sqlInterpolate method
DBI documentation: https://dbi.r-dbi.org/reference/dbBind.html

dbBind DBI method
Parametrised query is not officialy supported in Spark environment,
it is here managed with the sqlInterpolate method
DBI documentation: https://dbi.r-dbi.org/reference/dbBind.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
res <- dbGetQuery(db, "SELECT * FROM mtcars WHERE cyl == ?cyl")
dbBind(res, cyl = 4)
dbFetch(res)
}
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
res <- dbGetQuery(db, "SELECT * FROM mtcars WHERE cyl == ?cyl")
dbBind(res, cyl = 4)
dbFetch(res)
}
}
