% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbClearResult.R
\name{dbClearResult,SparkRResult-method}
\alias{dbClearResult,SparkRResult-method}
\title{dbClearResult DBI method
Release memory consumption of a SparkRResult object and prevent future Fetch
DBI documentation: https://dbi.r-dbi.org/reference/dbClearResult.html}
\usage{
\S4method{dbClearResult}{SparkRResult}(res, ...)
}
\description{
dbClearResult DBI method
Release memory consumption of a SparkRResult object and prevent future Fetch
DBI documentation: https://dbi.r-dbi.org/reference/dbClearResult.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
res <- dbGetQuery(db, "SELECT * FROM mtcars WHERE cyl == ?cyl")
dbClearResult(res)
}
}
