% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbColumnInfo.R
\name{dbColumnInfo,SparkRResult-method}
\alias{dbColumnInfo,SparkRResult-method}
\title{dbColumnInfo DBI method
Describe columns name, type and sql.type from a SparkRResult object
To correctly get the R type of the columns, the method evaluate and collect a row from the query. This can fail in case of the existence of heterogenous data type in the column (example: with NA an integer can be casted to float)
DBI documentation: https://dbi.r-dbi.org/reference/dbColumnInfo.html}
\usage{
\S4method{dbColumnInfo}{SparkRResult}(res, ...)
}
\description{
dbColumnInfo DBI method
Describe columns name, type and sql.type from a SparkRResult object
To correctly get the R type of the columns, the method evaluate and collect a row from the query. This can fail in case of the existence of heterogenous data type in the column (example: with NA an integer can be casted to float)
DBI documentation: https://dbi.r-dbi.org/reference/dbColumnInfo.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
res <- dbGetQuery(db, "SELECT * FROM mtcars WHERE cyl == ?cyl")
dbColumnInfo(res)
}
}
