% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCommit.R
\name{dbCommit,SparkRConnection-method}
\alias{dbCommit,SparkRConnection-method}
\title{dbBegin DBI method
Not supported in Spark environment
Defined in SparkR-DBI to keep compatibility with dbplyr
DBI documentation: https://dbi.r-dbi.org/reference/transactions.html}
\usage{
\S4method{dbCommit}{SparkRConnection}(conn)
}
\description{
dbBegin DBI method
Not supported in Spark environment
Defined in SparkR-DBI to keep compatibility with dbplyr
DBI documentation: https://dbi.r-dbi.org/reference/transactions.html
}
