% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbDataType.R
\name{dbDataType,SparkRConnection,character-method}
\alias{dbDataType,SparkRConnection,character-method}
\title{Find the database data type associated with an R object}
\usage{
\S4method{dbDataType}{SparkRConnection,character}(dbObj, obj, ...)
}
\description{
Find the database data type associated with an R object
}
