% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbDataType.R
\name{dbDataType,SparkRConnection-method}
\alias{dbDataType,SparkRConnection-method}
\title{dbDataType DBI method
Find the database data type related with an R object
This is performed by creating a Spark DF from the R object
and thus letting Spark automatically cast the object.
DBI documentation: https://dbi.r-dbi.org/reference/dbDataType.html}
\usage{
\S4method{dbDataType}{SparkRConnection}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{SparkRConnection object}

\item{obj}{R object}

\item{...}{Extra parameters}
}
\description{
dbDataType DBI method
Find the database data type related with an R object
This is performed by creating a Spark DF from the R object
and thus letting Spark automatically cast the object.
DBI documentation: https://dbi.r-dbi.org/reference/dbDataType.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbDataType(db, "test")
}
}
