% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbExistsTable.R
\name{dbExistsTable,SparkRConnection,ANY-method}
\alias{dbExistsTable,SparkRConnection,ANY-method}
\title{dbExistsTable DBI method
Check for table existence
If the table name if a simple string (example: abc), the table is searched inside the current used database.
If the table name contains a dot (example: databaseName.abc), the table is searched inside the specified database, named databaseName in this example.
DBI documentation: https://dbi.r-dbi.org/reference/dbExistsTable.html}
\usage{
\S4method{dbExistsTable}{SparkRConnection,ANY}(conn, name, ...)
}
\description{
dbExistsTable DBI method
Check for table existence
If the table name if a simple string (example: abc), the table is searched inside the current used database.
If the table name contains a dot (example: databaseName.abc), the table is searched inside the specified database, named databaseName in this example.
DBI documentation: https://dbi.r-dbi.org/reference/dbExistsTable.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
dbExistsTable(db, "mtcars")
}
}
