% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFetch.R
\name{dbFetch,SparkRResult-method}
\alias{dbFetch,SparkRResult-method}
\title{dbFetch DBI method
Fetch results from a SparkRResult object.
As Spark doesn't support cursor, we simulate this by fetching and keeping in memory the whole DataFrame during the first fetch and then deliver the requested subparts of that DataFrame.
DBI documentation: https://dbi.r-dbi.org/reference/dbFetch.html}
\usage{
\S4method{dbFetch}{SparkRResult}(res, n = -1, ...)
}
\description{
dbFetch DBI method
Fetch results from a SparkRResult object.
As Spark doesn't support cursor, we simulate this by fetching and keeping in memory the whole DataFrame during the first fetch and then deliver the requested subparts of that DataFrame.
DBI documentation: https://dbi.r-dbi.org/reference/dbFetch.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
res <- dbGetQuery(db, "SELECT * FROM mtcars")
dbFetch(res, n=10)
}
}
