% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetRowCount.R
\name{dbGetRowCount,SparkRResult-method}
\alias{dbGetRowCount,SparkRResult-method}
\title{dbGetRowCount DBI method
Return the returned number of rows from a SparkRResult object.
DBI documentation: https://dbi.r-dbi.org/reference/dbGetRowCount.html}
\usage{
\S4method{dbGetRowCount}{SparkRResult}(res, ...)
}
\description{
dbGetRowCount DBI method
Return the returned number of rows from a SparkRResult object.
DBI documentation: https://dbi.r-dbi.org/reference/dbGetRowCount.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
res <- dbGetQuery(db, "SELECT * FROM mtcars")
dbGetRowCount(res)
}
}
