% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetRowsAffected.R
\name{dbGetRowsAffected,SparkRResult-method}
\alias{dbGetRowsAffected,SparkRResult-method}
\title{dbGetRowsAffected DBI method
This functionnality is not supported by Spark.
We return 0 to keep compatibility with libraries using DBI object.
DBI documentation: https://dbi.r-dbi.org/reference/dbGetRowsAffected.html}
\usage{
\S4method{dbGetRowsAffected}{SparkRResult}(res, ...)
}
\description{
dbGetRowsAffected DBI method
This functionnality is not supported by Spark.
We return 0 to keep compatibility with libraries using DBI object.
DBI documentation: https://dbi.r-dbi.org/reference/dbGetRowsAffected.html
}
