% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetStatement.R
\name{dbGetStatement,SparkRResult-method}
\alias{dbGetStatement,SparkRResult-method}
\title{dbGetStatement DBI method
Return the statement of a SparkRResult object.
The returned statement is the providen one and not the statement generated after dbBind execution.
DBI documentation: https://dbi.r-dbi.org/reference/dbGetStatement.html}
\usage{
\S4method{dbGetStatement}{SparkRResult}(res, ...)
}
\arguments{
\item{res}{SparkRResult object}

\item{...}{Extra parameters}
}
\description{
dbGetStatement DBI method
Return the statement of a SparkRResult object.
The returned statement is the providen one and not the statement generated after dbBind execution.
DBI documentation: https://dbi.r-dbi.org/reference/dbGetStatement.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
res <- dbGetQuery(db, "SELECT * FROM mtcars")
dbGetStatement(res)
}
}
