% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbHasCompleted.R
\name{dbHasCompleted,SparkRResult-method}
\alias{dbHasCompleted,SparkRResult-method}
\title{dbHasCompleted DBI method
Return the completion status of a SparkRResult object.
DBI documentation: https://dbi.r-dbi.org/reference/dbHasCompleted.html}
\usage{
\S4method{dbHasCompleted}{SparkRResult}(res, ...)
}
\arguments{
\item{res}{SparkRResult object}

\item{...}{Extra parameters}
}
\description{
dbHasCompleted DBI method
Return the completion status of a SparkRResult object.
DBI documentation: https://dbi.r-dbi.org/reference/dbHasCompleted.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
res <- dbGetQuery(db, "SELECT * FROM mtcars")
dbFetch(res)
dbHasCompleted(res)
}
}
