% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbIsValid.R
\name{dbIsValid,SparkRConnection-method}
\alias{dbIsValid,SparkRConnection-method}
\title{dbIsValid DBI method
Return the valid status of the SparkR connection.
This method call the `isStopped` method of the sparkContext Scala API to get the connection status.
DBI documentation: https://dbi.r-dbi.org/reference/dbIsValid.html}
\usage{
\S4method{dbIsValid}{SparkRConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{SparkRConnection object}

\item{...}{Extra parameters}
}
\description{
dbIsValid DBI method
Return the valid status of the SparkR connection.
This method call the `isStopped` method of the sparkContext Scala API to get the connection status.
DBI documentation: https://dbi.r-dbi.org/reference/dbIsValid.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbIsValid(db)
}
}
