% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbListFields.R
\name{dbListFields,SparkRConnection,character-method}
\alias{dbListFields,SparkRConnection,character-method}
\title{dbListFields DBI method
Return the columns name according to a table name
(example: abc, databaseName.tableName, ...).
DBI documentation: https://dbi.r-dbi.org/reference/dbListFields.html}
\usage{
\S4method{dbListFields}{SparkRConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{SparkRConnection object}

\item{name}{Table name}

\item{...}{Extra parameters}
}
\description{
dbListFields DBI method
Return the columns name according to a table name
(example: abc, databaseName.tableName, ...).
DBI documentation: https://dbi.r-dbi.org/reference/dbListFields.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
dbListFields(db, "mtcars")
}
}
