% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbQuoteIdentifier.R
\name{dbQuoteIdentifier,SparkRConnection,character-method}
\alias{dbQuoteIdentifier,SparkRConnection,character-method}
\title{dbQuoteIdentifier DBI method
Generate quote to prevent SQL injection.
DBI documentation: https://dbi.r-dbi.org/reference/dbQuoteIdentifier.html}
\usage{
\S4method{dbQuoteIdentifier}{SparkRConnection,character}(conn, x)
}
\description{
dbQuoteIdentifier DBI method
Generate quote to prevent SQL injection.
DBI documentation: https://dbi.r-dbi.org/reference/dbQuoteIdentifier.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbQuoteIdentifier(db, "databaseName.columnName")
}
}
