% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSendQuery.R
\name{dbSendQuery,SparkRConnection-method}
\alias{dbSendQuery,SparkRConnection-method}
\title{Send a query to SparkR.}
\usage{
\S4method{dbSendQuery}{SparkRConnection}(conn, statement, ...)
}
\description{
Send a query to SparkR.
}
