% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbWriteTable.R
\name{dbWriteTable,SparkRConnection,ANY-method}
\alias{dbWriteTable,SparkRConnection,ANY-method}
\title{dbWriteTable DBI method
Write a table from a R dataframe.
DBI documentation: https://dbi.r-dbi.org/reference/dbWriteTable.html}
\usage{
\S4method{dbWriteTable}{SparkRConnection,ANY}(
  conn,
  name,
  value,
  append = FALSE,
  overwrite = FALSE,
  temporary = FALSE,
  row.names = FALSE,
  field.types = FALSE,
  ...
)
}
\description{
dbWriteTable DBI method
Write a table from a R dataframe.
DBI documentation: https://dbi.r-dbi.org/reference/dbWriteTable.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
}
}
