% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{class}
\name{SparkRConnection}
\alias{SparkRConnection}
\title{SparkR connection class}
\description{
SparkR connection class
}
\section{Slots}{

\describe{
\item{\code{sc}}{Spark Context object}
}}

\keyword{internal}
