% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.R
\docType{class}
\name{SparkRResult}
\alias{SparkRResult}
\title{SparkR results class.}
\description{
SparkR results class.
}
\section{Slots}{

\describe{
\item{\code{conn}}{SparkRConnection object}

\item{\code{state}}{Environment contaning the Result binded statement,
the fetching location (start and end), output data and
the cleared and completed state}

\item{\code{statement}}{Original provided statement (before binding)}
}}

\keyword{internal}
