% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetRowCount.R
\name{dbGetRowCount,SparkRResult-method}
\alias{dbGetRowCount,SparkRResult-method}
\title{Send a query to SparkR.}
\usage{
\S4method{dbGetRowCount}{SparkRResult}(res, ...)
}
\description{
Send a query to SparkR.
}
