% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRemoveTable.R
\name{dbRemoveTable,SparkRConnection,character-method}
\alias{dbRemoveTable,SparkRConnection,character-method}
\title{dbRemoveTable DBI method
Remove a table according to its name.
DBI documentation: https://dbi.r-dbi.org/reference/dbRemoveTable.html}
\usage{
\S4method{dbRemoveTable}{SparkRConnection,character}(conn, name, temporary = FALSE, fail_if_missing = TRUE, ...)
}
\description{
dbRemoveTable DBI method
Remove a table according to its name.
DBI documentation: https://dbi.r-dbi.org/reference/dbRemoveTable.html
}
\examples{
\dontrun{
db <- createSparkRConnection(sc=sc)
dbWriteTable(db, "mtcars", mtcars)
dbRemoveTable(db, "mtcars")
}
}
